requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cloudgenix==6.2.1b1
cryptography==41.0.1
future==0.18.3
html5lib==0.999999999
httpcache==0.1.3
idna==2.10
netmiko==4.2.0
ntc-templates==3.3.0
paramiko==3.2.0
Pillow==9.5.0
pycparser==2.21
pymal==0.6
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0
requests==2.25.1
scp==0.14.5
singleton-factory==1.0
singleton3==1.0
six==1.3.0
soupsieve==2.4.1
textfsm==1.1.3
urllib3==1.26.16
webencodings==0.5.1
websockets==11.0.3
